Move ScopeMeasure into a separate file
