+++ src/cache.cpp
@@ -16,6 +16,7 @@
 #include "logger.h"
 #include "matcherexception.h"
 #include "rssfeed.h"
+#include "scopemeasure.h"
 #include "strprintf.h"
 #include "utils.h"
