+++ src/rssfeed.cpp
@@ -19,6 +19,7 @@
 #include "dbexception.h"
 #include "htmlrenderer.h"
 #include "logger.h"
+#include "scopemeasure.h"
 #include "strprintf.h"
 #include "tagsouppullparser.h"
 #include "utils.h"
