+++ src/utils.cpp
@@ -572,42 +572,6 @@ unsigned int utils::to_u(const std::string& str,
 	return rs_to_u(str.c_str(), default_value);
 }
 
-ScopeMeasure::ScopeMeasure(const std::string& func, Level ll)
-	: funcname(func)
-	, lvl(ll)
-{
-	gettimeofday(&tv1, nullptr);
-}
-
-void ScopeMeasure::stopover(const std::string& son)
-{
-	gettimeofday(&tv2, nullptr);
-	unsigned long diff =
-		(((tv2.tv_sec - tv1.tv_sec) * 1000000) + tv2.tv_usec) -
-		tv1.tv_usec;
-	LOG(lvl,
-		"ScopeMeasure: function `%s' (stop over `%s') took %lu.%06lu "
-		"s so "
-		"far",
-		funcname,
-		son,
-		diff / 1000000,
-		diff % 1000000);
-}
-
-ScopeMeasure::~ScopeMeasure()
-{
-	gettimeofday(&tv2, nullptr);
-	unsigned long diff =
-		(((tv2.tv_sec - tv1.tv_sec) * 1000000) + tv2.tv_usec) -
-		tv1.tv_usec;
-	LOG(Level::INFO,
-		"ScopeMeasure: function `%s' took %lu.%06lu s",
-		funcname,
-		diff / 1000000,
-		diff % 1000000);
-}
-
 bool utils::is_valid_color(const std::string& color)
 {
 	return rs_is_valid_color(color.c_str());
